+++ src/protocol/mod.rs
@@ -8,6 +8,9 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use crate::errors::ApiError;
 
 pub mod e2e;
+mod message_id;
+
+pub use message_id::{MessageId, MessageIdParseError, MessageIds};
 
 /// A 16-byte blob ID.
 #[derive(Debug, PartialEq, Eq, Clone)]