+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Possible log types:
 - [added] Add a `RedisPublicKeyCache`. Enabled through the `public-key-cache-redis`
   Cargo feature.
 - [added] Add `E2eMessage` type
+- [added] Add `MessageId` type
 - [added] Add `E2eApi::decrypt_and_decode_incoming_message` method
 - [added] Add `E2eApi::encode_and_encrypt` method
 - [added] Derive `Clone`/`PartialEq` and implement `Deserialize` for FileMessage
@@ -33,6 +34,7 @@ Possible log types:
   module instead
 - [changed] The types `FileMessage`, `FileMessageBuilder` and `RenderingType` are not re-exported at top level
   anymore, import them from the `protocol::e2e::file` module instead
+- [changed] Change `IncomingMessage.message_id` type from `String` to `MessageId`
 - [deprecated] In `E2eApi` the `encrypt_text_msg`, `encrypt_image_msg`, `encrypt_file_msg` and `encrypt_raw`
   methods are deprecated and will be removed in a future release